Fix MSVC build https://bugs.webkit.org/show_bug.cgi?id=267017
Reviewed by Don Olmstead.

The default installation of MSVC is slightly different than what EWS is using.
These differences require a few small tweaks to get WebKit to compile.

* Source/WebCore/PAL/pal/text/TextCodec.cpp:
(PAL::TextCodec::getUnencodableReplacement):
%u in printf-like functions doesn't like char32_t in MSVC.  Just cast it to an unsigned.
* Source/WebKit/Scripts/generate-serializers.py:
(generate_impl):
IGNORE_WARNINGS_BEGIN("switch-unreachable") doesn't tell MSVC to ignore the right warning.
Instead, just make the generated code not have a default statement and exit the switch statement.
* Source/WebKit/Scripts/webkit/tests/GeneratedSerializers.cpp:
(WTF::void>):
Update the generated code tests along with the generator.
* Tools/TestWebKitAPI/Tests/WTF/DragonBoxTest.cpp:
(TestWebKitAPI::TEST):
MSVC needs constexpr to be able to allocate arrays on the stack.

Canonical link: https://commits.webkit.org/272595@main

+++ Source/WebCore/PAL/pal/text/TextCodec.cpp
@@ -46,9 +46,9 @@ int TextCodec::getUnencodableReplacement(char32_t codePoint, UnencodableHandling
 
     switch (handling) {
     case UnencodableHandling::Entities:
-        return snprintf(replacement.data(), sizeof(UnencodableReplacementArray), "&#%u;", codePoint);
+        return snprintf(replacement.data(), sizeof(UnencodableReplacementArray), "&#%u;", static_cast<unsigned>(codePoint));
     case UnencodableHandling::URLEncodedEntities:
-        return snprintf(replacement.data(), sizeof(UnencodableReplacementArray), "%%26%%23%u%%3B", codePoint);
+        return snprintf(replacement.data(), sizeof(UnencodableReplacementArray), "%%26%%23%u%%3B", static_cast<unsigned>(codePoint));
     }
     ASSERT_NOT_REACHED();
     replacement.data()[0] = 0;
